+++ src/main/java/graphEstructure/GraphInterface.java
@@ -16,7 +16,7 @@ public interface GraphInterface<T> {
 	public Iterable<Vertex<T>> getVertices();
 	public Edge<T> getEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
 	public Iterable<Edge<T>> getEdges(Vertex<T> vertex) throws IllegalArgumentException;
-	public double getWeightEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
+	public int getWeightEdge(Vertex<T> from, Vertex<T> to) throws IllegalArgumentException;
 	public void setWeightEdge(Vertex<T> from, Vertex<T> to, int weight) throws IllegalArgumentException;
 	public boolean isAdjacent(Vertex<T> from, Vertex<T> to);
 	public Set<Vertex<T>> adjacentVertices(Vertex<T> vertex);
